archive/submission_bundle_source/requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
opencv-python-headless>=4.8
tqdm>=4.66
